# Use swc for runtime bundling