@wyattjoh 2023-09-22
Current dependencies on/for this PR:
* canary
  * **PR #55842** <a href="https://app.graphite.dev/github/pr/vercel/next.js/55842" target="_blank"><img src="https://static.graphite.dev/graphite-32x32-black.png" alt="Graphite" width="10px" height="10px"/></a>  👈

This comment was auto-generated by [Graphite](https://app.graphite.dev/github/pr/vercel/next.js/55842?utm_source=stack-comment).